requirements.txt
albumentations==1.2.0
imgaug==0.4.0
joblib==1.0.1
millify==0.1.1
numpy==1.19.4
opencv_python==4.6.0.66
pandas==1.1.3
Pillow==9.2.0
pytest==7.1.2
scikit_learn==1.1.2
thop==0.1.1.post2207130030
timm==0.5.4
torch==1.8.1
torch_ema==0.3
torchvision==0.9.1
wandb==0.12.18